spada-dashboard/requirements.txt
playwright
playwright-stealth
python-dotenv
requests
fastapi
uvicorn[standard]